rename RedeemModal into RedeemNonModal
